Make install.py ask about deleting files

+++ install.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import os
+import shutil
 import subprocess
 
 BASE_DIR = os.path.abspath(os.path.dirname(__file__))
@@ -10,7 +11,29 @@ def install(dotfile, destination):
 
     print "installing %s -> %s..." % (source, destination)
     if os.path.lexists(destination):
-        os.remove(destination)
+        if not os.path.islink(destination):
+            while True:
+                delete = raw_input("%s (%s) exists and isn't a symlink, delete? (Y/n) " % \
+                                     (destination,
+                                      'directory' if os.path.isdir(destination) else 'file'))
+                delete = delete.strip().lower()
+                if delete == 'y' or not delete:
+                    delete = True
+                    break
+                elif delete == 'n':
+                    delete = False
+                    break
+        else:
+            delete = True
+
+        if not delete:
+            return
+
+        if os.path.islink(destination) or \
+           os.path.isfile(destination):
+            os.remove(destination)
+        else:
+            shutil.rmtree(destination)
 
     os.symlink(source, destination)
